packages/ai-personalization/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-generativeai==0.3.2
supabase
supabase-auth
python-dotenv==1.0.0
pandas==2.3.3
scikit-learn==1.4.0
numpy==1.26.3